+++ Changelog.md
@@ -1,5 +1,8 @@
 ## Ignition Transport 3.0.0
 
+1. Add ign topic --echo command line tool.
+    * [Pull request 189](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/189)
+
 1. Support a generic callback signature for receiving messages of any type.
     * [Pull request 188](https://bitbucket.org/ignitionrobotics/ign-transport/pull-request/188)